Refactored redundant check_aws_connection() helper functions into a single function in pge_utils.py
