+++ util/pge_util.py
@@ -105,6 +105,37 @@ def get_time_for_filename():
     return PRODUCTION_TIME
 
 
+def check_aws_connection(bucket, key):
+    """
+    Check connection to the provided S3 bucket by performing a test read
+    on the provided bucket/key location.
+
+    Parameters
+    ----------
+    bucket : str
+        Name of the S3 bucket to use with the connection test.
+    key : str, optional
+        S3 key path to append to the bucket name.
+
+    Raises
+    ------
+    RuntimeError
+        If not connection can be established.
+
+    """
+    s3 = boto3.resource('s3')
+    obj = s3.Object(bucket, key)
+
+    try:
+        logger.info(f'Attempting test read of s3://{obj.bucket_name}/{obj.key}')
+        obj.get()['Body'].read()
+        logger.info('Connection test successful.')
+    except Exception:
+        errmsg = (f'No access to the {bucket} S3 bucket. '
+                  f'Check your AWS credentials and re-run the code.')
+        raise RuntimeError(errmsg)
+
+
 def download_object_from_s3(s3_bucket, s3_key, output_filepath, filetype="Ancillary"):
     """Helper function to download an arbitrary file from S3"""
     if not s3_bucket or not s3_key:
